components: Add OpenTelemetry collector
OpenTelemetry collector is being added to monitoring-satellite as a router that receiver Jaeger traces from Gitpod components and redirects them to a Honeycomb dataset.

At the moment it is only added to our CI pipeline and preview environments, it will be enabled to production in a following PR after experimentation.

Signed-off-by: ArthurSens <arthursens2005@gmail.com>