[monitoring-central] remove psp
We're on K8s 1.27, it was removed from 1.25...at this point leaving it just makes shipping with ArgoCD harder (because it fails on PSP, and then we have to ignore when synchronizing).

The hope is that changling ths jsonnet file (barf), will cause CI to update the manifests (which are git ignored, but, include psp).